requirements.txt
greek-accentuation==0.9.9
inflexion==0.5
PyYAML==3.11
